[DC-9199] Incorporate review comment, create new controller for Bulk opt-outs as role has changed from admin to generic
