+++ conf/app.routes
@@ -26,8 +26,8 @@ POST        /message-brake/approve/confirmation         uk.gov.hmrc.preferencesa
 POST        /message-brake/reject/confirmation          uk.gov.hmrc.preferencesadminfrontend.controllers.MessageBrakeController.confirmRejectBatch()
 GET         /csv-upload                                 uk.gov.hmrc.preferencesadminfrontend.controllers.CsvUploadController.showUploadPage()
 POST        /csv-upload/confirmation                    uk.gov.hmrc.preferencesadminfrontend.controllers.CsvUploadController.upload()
-GET         /csv-upload-bulk-opt-outs                   uk.gov.hmrc.preferencesadminfrontend.controllers.CsvUploadController.showBulkOptOutsUploadPage
-POST        /csv-upload-bulk-opt-outs/confirmation      uk.gov.hmrc.preferencesadminfrontend.controllers.CsvUploadController.uploadBulkOptOuts()
+GET         /csv-upload-bulk-opt-outs                   uk.gov.hmrc.preferencesadminfrontend.controllers.CsvUploadBulkOptOutsController.showBulkOptOutsUploadPage
+POST        /csv-upload-bulk-opt-outs/confirmation      uk.gov.hmrc.preferencesadminfrontend.controllers.CsvUploadBulkOptOutsController.uploadBulkOptOuts()
 
 
 # Temporaty endpoits to migrate MDTP to ETMP
